Fixing potential error
Further testing showed that unlike how it was in my previous tests, that the first 18 elements are not always the same (and same position) and sent in every request.
In my personal test only the profile_background parameter was missing, so this code adds this parameter should it be missing in the correct place. I don't know if this is a one time problem or if it is needed to validate all of the 18 first elements of the form.