Improved behaviour of callbacks when a request to change a single item failed.
With the variable `showshowcaseconfigerrors` (true or false) in the values object it is now possible to control if the callback is called when a request fails. Additionally the callback is fired only once now in any case and the error message has the string appended, that the error happened when changing a specific item. (` | Happened while updating specific showcase items.`)