Rewrite from array to object.
Additionally now all values are unchanged if not supplied, or supplied as `undefined` in arrays. Showcases where the `values` property had only one value were changed to only this value instead of an additonaly object with f.e. "appid" as key. (Showcases were this changed are `game`,`review`,`group`). Additionally changed condition for errors in `ajaxsetshowcaseconfig` so they only throw an error in the correct scenario.