Removed leftover `this` from Promise, create new object instead of using reference
Additionally changing spacing around `if` clauses.

`Object.assign()` may not be a deep clone of the item object. But because only new properties are added directly to requestdata, the original item object is not modified.
(Only if item had a property like `item.prop = {"test": true}` and you would change `requestdata.prop.test = false` the property of item would be changed too, but if you would change f.e. `requestdata.prop = "test"` the item object would be unmodified, so I would say `Object.assign()` is okay for this case).