Update profile component to promises/update jsdoc
In the `/components/profile.js` file, all of the prototype functions
were refactored to use `callbackPromise`s from Doctor McKay's `StdLib`.
In the "for in" loops in `editProfile` and `profileSettings`,
the variable "i" was changed to "name" to be a bit more descriptive.
The JSDoc for each prototype function was also updated to reflect
the changes stated above.